seperated out ChosePhoneSwitches component
