+++ components/Onboarding/ChoosePhoneSwitches.js
@@ -1,0 +1,73 @@
+import React, { useState, useEffect } from 'react';
+import { Switch, Text, View } from 'react-native';
+
+import styles from '../../constants/screens/org-onboarding-styles/ChoosePhoneSwitches';
+
+const ChoosePhoneSwitches = props => {
+  const [applePhone, setApplePhone] = useState(false);
+  const [androidPhone, setAndroidPhone] = useState(false);
+  const [otherPhone, setOtherPhone] = useState(false);
+
+  useEffect(() => {
+    let smartphoneType = '';
+    const types = ['Apple', 'Android', 'Other'];
+    const bools = [applePhone, androidPhone, otherPhone];
+    types.forEach((type, index) => {
+      if (bools[index]) {
+        if (smartphoneType.length) {
+          smartphoneType = `${smartphoneType}, ${type}`;
+        } else smartphoneType = type;
+      }
+    });
+    props.onChangeText({
+      ...props.airtableState,
+      smartphone_type: smartphoneType
+    });
+  }, [
+    applePhone,
+    androidPhone,
+    otherPhone,
+    props.airtableState.smartphone_type
+  ]);
+
+  return (
+    <React.Fragment>
+      <Text style={styles.obText}>
+        What type of smartphones do you use?
+        <Text style={[styles.obText, styles.italic]}>
+          {' '}
+          Select All that apply.
+        </Text>
+      </Text>
+      <View style={styles.switchContainer}>
+        <Switch
+          trackColor={{ true: '#00FF9D' }}
+          style={styles.obSwitchButton}
+          value={applePhone}
+          onValueChange={newValue => setApplePhone(newValue)}
+        />
+        <Text style={styles.obSwitchLabel}>Apple</Text>
+      </View>
+      <View style={styles.switchContainer}>
+        <Switch
+          trackColor={{ true: '#00FF9D' }}
+          style={styles.obSwitchButton}
+          value={androidPhone}
+          onValueChange={newValue => setAndroidPhone(newValue)}
+        />
+        <Text style={styles.obSwitchLabel}>Android</Text>
+      </View>
+      <View style={styles.switchContainer}>
+        <Switch
+          trackColor={{ true: '#00FF9D' }}
+          style={styles.obSwitchButton}
+          value={otherPhone}
+          onValueChange={newValue => setOtherPhone(newValue)}
+        />
+        <Text style={styles.obSwitchLabel}>Other</Text>
+      </View>
+    </React.Fragment>
+  );
+};
+
+export default ChoosePhoneSwitches;
